feat: Characters route
Added a route in which characters can be added, deleted, updated from the characters table. Also added routes to get all characters within a lobby, add characters to a lobby, ability to grant characters to a user within a lobby. Ability to see what characters a user has within a lobby.